+++ py-scouter/python/scouter/queue/__init__.pyi
@@ -1,8 +1,11 @@
-from typing import Dict, Optional, Union
+from typing import Dict, Optional, Union, List, Any
+import datetime
 
-from .. import Features, PsiDriftProfile, ServerRecords, SpcDriftProfile
+from ..drift import PsiDriftProfile, SpcDriftProfile
 from ..client import HTTPConfig
 from ..logging import LogLevel
+from ..types import RecordType
+from ..observe import ObservabilityMetrics
 
 class KafkaConfig:
     brokers: str
@@ -66,8 +69,6 @@ class KafkaConfig:
                 Default is 3.
 
         """
-
-        ...
 
 class RabbitMQConfig:
     address: str
@@ -133,8 +134,6 @@ class ScouterProducer:
                 Default is 3.
         """
 
-        ...
-
     def publish(self, message: ServerRecords) -> None:
         """Publish a message to the queue.
 
@@ -143,12 +142,8 @@ class ScouterProducer:
                 Message to publish.
         """
 
-        ...
-
     def flush(self) -> None:
         """Flush the producer queue."""
-
-        ...
 
 class ScouterQueue:
     def __init__(
@@ -170,8 +165,6 @@ class ScouterQueue:
                 Default is 3.
         """
 
-        ...
-
     def insert(self, features: Features) -> None:
         """Insert features into the queue.
 
@@ -180,9 +173,338 @@ class ScouterQueue:
                 Features to insert.
         """
 
-        ...
-
     def flush(self) -> None:
         """Flush the queue."""
 
-        ...
+
+class SpcServerRecord:
+    def __init__(
+        self,
+        repository: str,
+        name: str,
+        version: str,
+        feature: str,
+        value: float,
+    ):
+        """Initialize spc drift server record
+
+        Args:
+            repository:
+                Model repository
+            name:
+                Model name
+            version:
+                Model version
+            feature:
+                Feature name
+            value:
+                Feature value
+        """
+
+    @property
+    def created_at(self) -> datetime.datetime:
+        """Return the created at timestamp."""
+
+    @property
+    def repository(self) -> str:
+        """Return the repository."""
+
+    @property
+    def name(self) -> str:
+        """Return the name."""
+
+    @property
+    def version(self) -> str:
+        """Return the version."""
+
+    @property
+    def feature(self) -> str:
+        """Return the feature."""
+
+    @property
+    def value(self) -> float:
+        """Return the sample value."""
+
+    def __str__(self) -> str:
+        """Return the string representation of the record."""
+
+    def model_dump_json(self) -> str:
+        """Return the json representation of the record."""
+
+    def to_dict(self) -> Dict[str, str]:
+        """Return the dictionary representation of the record."""
+
+class PsiServerRecord:
+    def __init__(
+        self,
+        repository: str,
+        name: str,
+        version: str,
+        feature: str,
+        bin_id: str,
+        bin_count: int,
+    ):
+        """Initialize spc drift server record
+
+        Args:
+            repository:
+                Model repository
+            name:
+                Model name
+            version:
+                Model version
+            feature:
+                Feature name
+            bin_id:
+                Bundle ID
+            bin_count:
+                Bundle ID
+        """
+
+    @property
+    def created_at(self) -> datetime.datetime:
+        """Return the created at timestamp."""
+
+    @property
+    def repository(self) -> str:
+        """Return the repository."""
+
+    @property
+    def name(self) -> str:
+        """Return the name."""
+
+    @property
+    def version(self) -> str:
+        """Return the version."""
+
+    @property
+    def feature(self) -> str:
+        """Return the feature."""
+
+    @property
+    def bin_id(self) -> str:
+        """Return the sample value."""
+
+    @property
+    def bin_count(self) -> int:
+        """Return the sample value."""
+
+    def __str__(self) -> str:
+        """Return the string representation of the record."""
+
+    def model_dump_json(self) -> str:
+        """Return the json representation of the record."""
+
+    def to_dict(self) -> Dict[str, str]:
+        """Return the dictionary representation of the record."""
+    
+
+class CustomMetricServerRecord:
+    def __init__(
+        self,
+        repository: str,
+        name: str,
+        version: str,
+        metric: str,
+        value: int,
+    ):
+        """Initialize spc drift server record
+
+        Args:
+            repository:
+                Model repository
+            name:
+                Model name
+            version:
+                Model version
+            metric:
+                Metric name
+            value:
+                Metric value
+        """
+
+    @property
+    def created_at(self) -> datetime.datetime:
+        """Return the created at timestamp."""
+
+    @property
+    def repository(self) -> str:
+        """Return the repository."""
+
+    @property
+    def name(self) -> str:
+        """Return the name."""
+
+    @property
+    def version(self) -> str:
+        """Return the version."""
+
+    @property
+    def metric(self) -> str:
+        """Return the metric name."""
+
+    @property
+    def value(self) -> float:
+        """Return the metric value."""
+
+    def __str__(self) -> str:
+        """Return the string representation of the record."""
+
+    def model_dump_json(self) -> str:
+        """Return the json representation of the record."""
+
+    def to_dict(self) -> Dict[str, str]:
+        """Return the dictionary representation of the record."""
+
+
+class ServerRecord:
+    Spc: "ServerRecord"
+    Psi: "ServerRecord"
+    Custom: "ServerRecord"
+    Observability: "ServerRecord"
+
+    def __init__(self, record: Any) -> None:
+        """Initialize server record
+
+        Args:
+            record:
+                Server record to initialize
+        """
+
+    @property
+    def record(self) -> Union[SpcServerRecord, PsiServerRecord, CustomMetricServerRecord, ObservabilityMetrics]:
+        """Return the drift server record."""
+
+class ServerRecords:
+    def __init__(self, records: List[ServerRecord], record_type: RecordType) -> None:
+        """Initialize server records
+
+        Args:
+            records:
+                List of server records
+            record_type:
+                Type of server records
+        """
+
+    @property
+    def record_type(self) -> RecordType:
+        """Return the drift type."""
+
+    @property
+    def records(self) -> List[ServerRecord]:
+        """Return the drift server records."""
+
+    def model_dump_json(self) -> str:
+        """Return the json representation of the record."""
+
+    def __str__(self) -> str:
+        """Return the string representation of the record."""
+
+
+
+class Feature:
+    @staticmethod
+    def int(name: str, value: int) -> "Feature":
+        """Create an integer feature
+
+        Args:
+            name:
+                Name of the feature
+            value:
+                Value of the feature
+        """
+
+    @staticmethod
+    def float(name: str, value: float) -> "Feature":
+        """Create a float feature
+
+        Args:
+            name:
+                Name of the feature
+            value:
+                Value of the feature
+        """
+
+    @staticmethod
+    def string(name: str, value: str) -> "Feature":
+        """Create a string feature
+
+        Args:
+            name:
+                Name of the feature
+            value:
+                Value of the feature
+        """
+
+class Features:
+    def __init__(self, features: List[Feature]) -> None:
+        """Initialize features
+
+        Args:
+            features:
+                List of features
+        """
+
+    def __str__(self) -> str:
+        """Return the string representation of the features"""
+
+class PsiFeatureQueue:
+    def __init__(self, drift_profile: PsiDriftProfile) -> None:
+        """Initialize the feature queue
+
+        Args:
+            drift_profile:
+                Drift profile to use for feature queue.
+        """
+
+    def insert(self, features: Features) -> None:
+        """Insert data into the feature queue
+        Args:
+            features:
+                List of features to insert into the monitoring queue.
+        """
+
+    def is_empty(self) -> bool:
+        """check if queue is empty
+        Returns:
+            bool
+        """
+
+    def clear_queue(self) -> None:
+        """Clears the feature queue"""
+
+    def create_drift_records(self) -> ServerRecords:
+        """Create drift server record from data
+
+
+        Returns:
+            `DriftServerRecord`
+        """
+
+class SpcFeatureQueue:
+    def __init__(self, drift_profile: SpcDriftProfile) -> None:
+        """Initialize the feature queue
+
+        Args:
+            drift_profile:
+                Drift profile to use for feature queue.
+        """
+
+    def insert(self, features: Features) -> None:
+        """Insert data into the feature queue
+
+        Args:
+            features:
+                List of features to insert into the monitoring queue.
+        """
+
+    def create_drift_records(self) -> ServerRecords:
+        """Create drift server record from data
+
+
+        Returns:
+            `DriftServerRecord`
+        """
+
+    def clear_queue(self) -> None:
+        """Clears the feature queue"""
